refactor(cli): split component history into focused files
Extract the rendering, customization-collection, and git-metrics logic out of the ~1100-line history.go into history_render.go, history_customizations.go, and history_gitmetrics.go. All files remain in package component; no behavior change.
